refactor!: move sequence logic out of api.py
