+++ src/aioamazondevices/implementation/__init__.py
@@ -1,0 +1,1 @@
+"""aioamazondevices implementation package."""
